// src/interfaces/refs.ts
export interface Refs {
  nameRef?: React.RefObject<HTMLInputElement>;
  dateRef?: React.RefObject<HTMLInputElement>;
  typeRef?: React.RefObject<HTMLSelectElement>;
  costRef?: React.RefObject<HTMLInputElement>;
  usdtRef?: React.RefObject<HTMLInputElement>;
  ethRef?: React.RefObject<HTMLInputElement>;
  fileRef?: React.RefObject<HTMLInputElement>;
  agreeRef?: React.RefObject<HTMLInputElement>;
}